test: LikedCafe 분리에 따른 테스트 수정
#236
